Single-consumer A4a: process-unique 64-bit correlation id (no batch_id aliasing)
Widen CORRELATION_ID_TYPE from 32 to 64 bits and repack the networked consumer's
wire token as a 48-bit batch id + 16-bit slot. The batch id is drawn from one
monotonic process-wide counter on the shared consumer, so concurrent submitters
never collide; at 48 bits it does not wrap in any realistic run. This removes the
silent-misrouting hazard that the previous 16-bit batch id allowed -- once two
algorithms share one networked consumer, a stale/late return from a retired batch
could otherwise alias a freshly-minted batch with the same wrapped id and be
checked in to the wrong submitter's slot.

With non-wrapping ids the late-return buffer's TTL no longer needs to guard against
batch_id reuse, so its wraparound-coupling clamp and the related comments are
removed (the epoch tag is now used purely for TTL eviction).

Release build + ctest 59/59 (repeatable); MPI ex16 -np 4 converges with the wider
wire token.

+++ courtier/include/courtier/GCourtierEnums.hpp
@@ -278,7 +278,10 @@ constexpr std::size_t DEFAULTMAXRESUBMISSIONS = 5;
 using ITERATION_COUNTER_TYPE = std::uint64_t;
 using RESUBMISSION_COUNTER_TYPE = std::size_t;
 using COLLECTION_POSITION_TYPE = std::size_t;
-using CORRELATION_ID_TYPE = std::uint32_t;
+// 64-bit so the networked consumer can pack a 48-bit, process-unique, never-wrapping batch id with a
+// 16-bit slot index (GNetworkedConsumerT). A wide batch id is what makes late-return routing safe when
+// many algorithms submit through one shared consumer: a stale return cannot alias a freshly-minted batch.
+using CORRELATION_ID_TYPE = std::uint64_t;
 
 /******************************************************************************/
